Code appears to be working for most cases. extrapolated commonly used code to utils.py updated bound checking in Types.py to be more readable and better debugable
