+++ Helpers.py
@@ -2,12 +2,7 @@ import json
 import os
 from base64 import b64encode
 
-import cv2
 import requests
-import time
-
-import Config
-from Types import TextField
 
 
 class GCloudOCR:
@@ -48,45 +43,4 @@ class GCloudOCR:
                              )
 
 
-def in_range(value: int, min: int, max: int):
-    return min <= int <= max
 
-
-def debug_condition(p1: int, pc1: str, p2: int, pc2: str, p3: int):
-    print("%d %s %d %s %d " % (p1, pc1, p2, pc2, p3))
-
-
-def json_to_field(list):
-    field_list = []
-    for field in list:
-        field_list.append(TextField(field, "Generic"))
-    return field_list
-
-
-def display_fields(image, fields, colour, show_text=False):
-    bounds = None
-    for field in fields:
-        bounds = field.get_bounds()
-        try:
-            cv2.polylines(image, np.int32([np.array(bounds.get_as_points())]), 1, colour)
-            if show_text:
-                cv2.putText(image, field.get_value() + "(" + field.get_field_type() + ")",
-                            (bounds.br('x') + 5, bounds.br('y')),
-                            cv2.FONT_HERSHEY_COMPLEX_SMALL, .5, Config.Colors.outline, 3)
-                cv2.putText(image, field.get_value() + "(" + field.get_field_type() + ")",
-                            (bounds.br('x') + 5, bounds.br('y')),
-                            cv2.FONT_HERSHEY_COMPLEX_SMALL, .5, Config.Colors.neutral, 1)
-        except ValueError:
-            print("Value Error: " + field)
-        except TypeError:
-            print(bounds)
-
-
-def insert_record(output_csv, card) -> None:
-    now = time.localtime()
-    snumber = card.get_student_id()
-    name = card.name_as_str()
-    output_csv.write(
-        "s%s,%s,s%s@student.rmit.edu.au,%s\n" % (snumber, name, snumber, "%.4d/%.2d/%.2d %.2d:%.2d" % (
-            now.tm_year, now.tm_mon, now.tm_mday, now.tm_hour, now.tm_min)))
-    output_csv.flush()
